requirements.txt
absl-py==0.10.0
appdirs==1.4.4
asgiref==3.4.1
astroid==2.3.3
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
autopep8==1.5.7
awsebcli==3.20.3
backcall==0.1.0
bleach==3.1.4
botocore==1.23.54
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.8
django-extensions==3.1.5
django-on-heroku==1.1.2
entrypoints==0.3
filelock==3.0.12
future==0.16.0
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imageio==2.8.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server==0.1.1
jupyterlab-pygments==0.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
onnx==1.7.0
onnx-tf==1.5.0
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
opt-einsum==3.3.0
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.13.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pypiwin32==223
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.16.2
scipy==1.4.1
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.4.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.9
virtualenv==20.0.31
voila==0.1.21
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==6.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
